pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supcon"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1",
  "torchvision>=0.16",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "wandb>=0.16",
  "tqdm>=4.66",
  "numpy>=1.24",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
